[Fix] Syntax fix. Bash globbing does not accept spaces between var="string"
Syntax error presents if spaces are present between variable and it's newly set value. Since firewall pipes things to /dev/null we haven't seen any errors, until we noticed the value was not being properly updated by APF. Also if the variable is commented out in conf.apf this will not trigger the setting as it is unset versus empty, so an additional or is provided empty or unset.